chore: rename dtos
