+++ api/src/users/auth.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test } from '@nestjs/testing';
 import { ObjectID } from 'typeorm';
 import { AuthService } from './auth.service';
-import { CreateUserDto } from './dtos/create-user-dto';
+import { CreateUserDto } from './dtos/create-user.dto';
 import { User } from './user.entity';
 import { UsersService } from './users.service';
